Fixed REB3-12048: RuntimeError incompatible numpy version
DetectMultiBackend runs `pip install onnxruntime` in case when `onnxruntime-gpu` is present in the system and CUDA is not available. This is actually not required as `onnxruntime-gpu` already contains all the needed providers.

It's a bad practice to install packages dynamically during runtime ignoring the requirements that developers are tailored for the execution environment.
This leads to a binary incompatibility of the packages and other issues. And also in some cases it could just slow down system cold starts.
It's better to check whether system initialization is possible in the current environment and fallback to package installation if there's no other way to work.